+++ package.json
@@ -12,6 +12,8 @@
   "dependencies": {
     "@expo-google-fonts/roboto": "^0.2.2",
     "@react-native-masked-view/masked-view": "0.2.6",
+    "@react-navigation/bottom-tabs": "^6.3.1",
+    "@react-navigation/material-bottom-tabs": "^6.2.1",
     "@react-navigation/native": "^6.0.10",
     "@react-navigation/stack": "^6.1.1",
     "axios": "^0.26.0",
@@ -22,15 +24,19 @@
     "expo-linear-gradient": "~11.0.3",
     "expo-random": "~12.1.1",
     "expo-status-bar": "~1.2.0",
+    "feather-icons": "^4.29.0",
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-native": "0.64.3",
     "react-native-gesture-handler": "~2.1.0",
+    "react-native-ionicons": "^4.6.5",
     "react-native-iphone-x-helper": "^1.3.1",
+    "react-native-reanimated": "~2.3.1",
     "react-native-safe-area-context": "3.3.2",
     "react-native-screens": "~3.10.1",
     "react-native-svg": "12.1.1",
-    "react-native-web": "0.17.1"
+    "react-native-web": "0.17.1",
+    "react-navigation": "^4.4.4"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
